--- crt_portal/api/tests/test_data.py
@@ -1,29 +1,0 @@
-SAMPLE_ACTION_1 = {
-    'id': 1,
-    'actor_object_id': 1,
-    'verb': 'Contacted complainant:',
-    'description': "Email sent: 'EOS - Department of Ed OCR Referral Form Letter' to cookiemonster@fakeemail.com via govDelivery TMS",
-    'timestamp': '2022-04-12 14:56:53.277142+00',
-    'public': True,
-    'actor_content_type_id': 1
-}
-
-SAMPLE_ACTION_2 = {
-    'id': 2,
-    'actor_object_id': 1,
-    'verb': 'Contacted complainant:',
-    'description': "Email sent: 'EOS - EEOC Referral Form Letter' to    eileenmcfarland@navapbc.com via govDelivery TMS",
-    'timestamp': '2022-04-12 17:30:53.277142+00',
-    'public': True,
-    'actor_content_type_id': 1
-}
-
-SAMPLE_ACTION_3 = {
-    'id': 3,
-    'actor_object_id': 1,
-    'verb': 'Contacted complainant:',
-    'description': "Email sent: 'EOS - EEOC Referral Form Letter' to  eileenmcfarland@navapbc.com via govDelivery TMS",
-    'timestamp': '2022-04-15 10:56:53.277142+00',
-    'public': True,
-    'actor_content_type_id': 1
-}
